crm_core/dev_requirements.txt
Django
djangorestframework
python-dotenv
amqp
redis
celery
flower
pymongo dnspython
